Add the initial rough work for the emulator
This commit is large. It introduces the cpu.rs module which handles interpreting the 8080 opcodes and translating them to Rust code. I also made a new script (generate_op_funcs.py) which generates the boilerplate for the match statement in Cpu::execute().

This version is rough, but it's a start. As of now the program should be run like so `RUST_LOG=debug cargo run -- invaders.rom`