Introduce math functions and fix stack operations
The math functions exist as a way to set flags automatically, they need reworking, and I'll do that when I have a better picture of the requirements.

The stack was causing issues due to missing parenthesis around one of the bitwise shifts, the issue has been fixed and CALL/RET functionality works as intended now :).

Newly implemented opcodes:
   0x05 DCR B
   0x13 INX D
   0x1a LDAX D
   0x23 INX H
   0x77 MOV M,A
   0xc9 RET
Modified opcodes:
   0xcd CALL adr -- Fix missing parenthesis causing abnormal behaviour; Also advance the PC stored on the stack so the RET doesn't infinitely loop.